requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-ollama==0.3.7
langchain-core==0.3.75
langchain-deepseek==0.1.4
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.31.0